requirements.txt
pandas
streamlit
plotly
scikit-learn
ollama
python-dotenv
openpyxl